pkg: operators: wasm: Move syscall_helpers functions to pkg/utils/syscalls.
Before 3ba4e5500d1a ("treewide: Remove built-in gadgets.") this package was
shared with builtin traceloop.
This no more makes sense today, as builtins were removed.
So, let's move this file to proper location.
